Connect simulation table
Now, entries in the SimulationTableModel contain enough info to
configure a simulation run. Simulations are configured based on these
entry objects. The result ("score") is saved and displayed.

Up next: allow multiple configurations, and make the configurations
configurable.

+++ fahbench/gui/SimulationWorker.cpp
@@ -28,7 +28,7 @@ SimulationWorker::SimulationWorker(): QObject() {
 
 void SimulationWorker::run_simulation(Simulation simulation) {
     try {
-        double score = simulation.run(*this);
+        float score = simulation.run(*this);
         emit simulation_finished(score);
     } catch (std::exception & e) {
         message(e.what());
